Alire build support (#25)
* Buildable using Alire

alire.toml derived from the one in the community index. Added project
files, additional tags and version updated to 1.1.0-dev.

Project files are not yet using Alite capabilities.

* Update protobuf_runtime.gpr to standard Alire 1.2.1 configuration

Included config directory in git in order to support building without
Alire.

* More integration with Alire

- conformance.gpr and protoc_gen_ada.gpr use now Alire config GPR.
- Separate object dirs for the different build profiles as used by Alire
  by default.
- Restore "library project" for readability.

* Align version in rpm to Alire

* Update the pre-release part of the Alire version to match the RPM

The pre-release part of the Alire version is the same as the fixed part of
the release identifier in the RPM. It is added to the library filename so
it matches the file generated using the crate version.

* Fix double definition of Library_Kind attribute

And set default to relocatable.